requirements.txt
Django==5.2.4
Pillow==11.2.1
django-mptt==0.17.0
django-allauth==0.57.0
django-ckeditor==6.7.0
python-decouple==3.8
django-cors-headers==4.3.1
django-environ==0.11.2
requests==2.31.0
whitenoise==6.6.0
gunicorn==23.0.0
django-extensions==3.2.3
django-debug-toolbar==4.2.0
celery==5.3.4
redis==5.0.1
django-countries==7.5.1
django-phonenumber-field==7.1.0
phonenumbers==8.13.26
pymysql==1.1.1
mysqlclient==2.2.7
djangorestframework==3.14.0
drf-yasg==1.21.7
django-filter==23.5
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.27.0